+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [5.0.0] - 2026-06-02
+
+_Stable release based on [5.0.0-rc.1]._
+
 ## [5.0.0-rc.1] - 2026-06-02
 
 ### Changed
@@ -713,6 +717,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 - New changelog file.
 
+[5.0.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v4.2.1...v5.0.0
 [5.0.0-rc.1]: https://github.com/internetguru/laravel-model-browser/releases/tag/v4.2.1
 [4.2.1]: https://https://github.com/internetguru/laravel-model-browser/compare/v4.2.0...v4.2.1
 [4.2.0]: https://https://github.com/internetguru/laravel-model-browser/compare/v4.1.7...v4.2.0